docker/sdf_converter/requirements.txt
flask==2.3.0
gunicorn==21.2.0
google-cloud-storage==2.10.0